get it working for a Wire
