+++ crates/opencascade/src/primitives/face.rs
@@ -1,13 +1,12 @@
 use crate::{
     angle::Angle,
     primitives::{
-        make_axis_1, make_point, make_point2d, make_vec, EdgeIterator, JoinType, Shape, Solid,
-        Surface, Wire,
+        make_axis_1, make_point, make_vec, EdgeIterator, JoinType, Shape, Solid, Surface, Wire,
     },
     workplane::Workplane,
 };
 use cxx::UniquePtr;
-use glam::{dvec2, dvec3, DVec3};
+use glam::{dvec3, DVec3};
 use opencascade_sys::ffi;
 
 pub struct Face {
@@ -209,41 +208,6 @@ impl Face {
         let mut make_pipe = ffi::BRepOffsetAPI_MakePipe_ctor(&path.inner, profile_shape);
 
         let pipe_shape = make_pipe.pin_mut().Shape();
-        let result_solid = ffi::TopoDS_cast_to_solid(pipe_shape);
-
-        Solid::from_solid(result_solid)
-    }
-
-    /// Sweep the face along a path, modulated by a function, to produce a solid
-    #[must_use]
-    pub fn sweep_along_with_radius_values(
-        &self,
-        path: &Wire,
-        radius_values: impl IntoIterator<Item = (f64, f64)>,
-    ) -> Solid {
-        // let radius_values: Vec<_> = radius_values.into_iter().collect();
-        // let mut array = ffi::TColgp_Array1OfPnt2d_ctor(1, radius_values.len() as i32);
-
-        // for (index, (t, radius)) in radius_values.into_iter().enumerate() {
-        //     array.pin_mut().SetValue(index as i32 + 1, &make_point2d(dvec2(t, radius)));
-        // }
-
-        // let mut interpol = ffi::Law_Interpol_ctor();
-        // let is_periodic = false;
-        // interpol.pin_mut().Set(&array, is_periodic);
-        // let law_function = ffi::Law_Interpol_into_Law_Function(interpol);
-        // let law_handle = ffi::Law_Function_to_handle(law_function);
-
-        let profile_shape = ffi::cast_face_to_shape(&self.inner);
-        let mut make_pipe_shell = ffi::BRepOffsetAPI_MakePipeShell_ctor(&path.inner);
-        let with_contact = false;
-        let with_correction = true;
-        // make_pipe_shell.pin_mut().SetLaw(profile_shape, &law_handle, with_contact, with_correction);
-        println!("1 POTATO");
-        make_pipe_shell.pin_mut().Add(profile_shape, with_contact, with_correction);
-        println!("2 POTATO");
-
-        let pipe_shape = make_pipe_shell.pin_mut().Shape();
         let result_solid = ffi::TopoDS_cast_to_solid(pipe_shape);
 
         Solid::from_solid(result_solid)
