+++ examples/src/swept_wire_variable.rs
@@ -1,7 +1,7 @@
 use glam::DVec3;
 use opencascade::{
     angle::{RVec, ToAngle},
-    primitives::{approximate_function, Face, IntoShape, Shape, Solid, Wire},
+    primitives::{approximate_function, IntoShape, Shape, Shell, Wire},
     workplane::Workplane,
 };
 
@@ -9,16 +9,15 @@ pub fn shape() -> Shape {
     let r = 10.0;
     let a = 5.0;
 
-    let face_profile: Face = Workplane::xz()
+    let wire_profile: Wire = Workplane::xz()
         .rotated(RVec::z(45.0.degrees()))
         .translated(DVec3::new(-r, 0.0, 0.0))
-        .rect(a, a)
-        .to_face();
+        .rect(a, a);
 
     let path: Wire = Workplane::xy().sketch().arc((-r, 0.0), (0.0, r), (r, 0.0)).wire();
 
     let num_radii = 5;
-    let pipe_solid: Solid = face_profile.sweep_along_with_radius_values(
+    let pipe_shell: Shell = wire_profile.sweep_along_with_radius_values(
         &path,
         approximate_function(num_radii, |t| {
             let val = ((2.0 * std::f64::consts::PI * (t - 1.0 / 4.0)).sin() + 1.0) / 2.0;
@@ -26,5 +25,5 @@ pub fn shape() -> Shape {
         }),
     );
 
-    pipe_solid.into_shape()
+    pipe_shell.into_shape()
 }
